Utils refactor
