--- src/utils/findWorkspace.ts
@@ -1,18 +1,0 @@
-import * as path from 'path'
-import { isWorkspacePath } from './isWorkspacePath'
-
-export const findWorkspace = async (
-  start: string
-): Promise<string | undefined> => {
-  if (await isWorkspacePath(start)) {
-    return start
-  }
-
-  const parent = path.dirname(start)
-
-  if (!parent || parent === '/' || parent === '.') {
-    return undefined
-  }
-
-  return findWorkspace(parent)
-}
