--- src/utils/silentReporter.ts
@@ -1,8 +1,0 @@
-import { Reporter } from './reporter'
-
-export const silentReporter: Reporter = {
-  log: () => {},
-  info: () => {},
-  warn: () => {},
-  error: () => {},
-}
